Decompose state send scheduling out to a separate generator
